Rename/Move Settings.ts -> settings/SettingsTab.ts
