+++ src/settings/SettingsTab.ts
@@ -1,8 +1,8 @@
 import { App, Notice, PluginSettingTab, Setting, normalizePath } from 'obsidian';
-import SettingsProfilesPlugin from './main';
-import { DEFAULT_PROFILE_SETTINGS } from './interface';
-import { loadProfileData } from './util/SettingsFiles';
-import { ProfileSettingsModal } from './modals/ProfileSettingsModal';
+import SettingsProfilesPlugin from '../main';
+import { DEFAULT_PROFILE_SETTINGS } from '../interface';
+import { loadProfileData } from '../util/SettingsFiles';
+import { ProfileSettingsModal } from '../modals/ProfileSettingsModal';
 export class SettingsProfilesSettingTab extends PluginSettingTab {
 	plugin: SettingsProfilesPlugin;
 	profilesSettings: Setting[];
